Cambios en mensajes de la aplicacion, algun campo erroneo y mas cambios visuales, cambio de version

+++ app/src/main/java/com/example/padeldam/back/dao/PistaRepositorio.java
@@ -36,7 +36,7 @@ public class PistaRepositorio implements IPistas<Pista> {
                                             document.get("nombre", String.class),
                                             document.get("numero", Integer.class),
                                             document.get("material", String.class),
-                                            document.get("precioHora", Integer.class)
+                                            document.get("precioHora", Double.class)
 
                                     );
                                     pistas.add(p);
